refactor: 컴포넌트 관심사 및 상태 분리를 위해 마지막 로그인 계정 관리하는 컴포넌트 생성
